requirements.txt
albumentations
av2
muse_maskgit_pytorch
scipy
numpy
matplotlib
tqdm
opencv-python
pandas
pillow
einops
nuscenes-devkit
av2
hydra-core==1.2.0
hydra_colorlog
torch-tb-profiler
tensorboard
deepspeed
pytorch-lightning<2.0.0
wandb>=0.13.10
rich
pyrootutils
triton>=2.0.0
clean-fid
typer
metadrive-simulator
torchmetrics[image]
--extra-index-url https://download.pytorch.org/whl/cu117
torch<=1.13.1
torchvision<=0.14.1
git+https://github.com/alexanderswerdlow/image_utils.git
git+https://github.com/alexanderswerdlow/lightning-bolts.git@fix-lightningloggerbase